+++ spec/dummy_environments/sqlite/dummy1/models/author.rb
@@ -1,0 +1,5 @@
+class Author < ActiveRecord::Base
+  attr_accessible :name
+  has_many :posts
+  has_many :comments
+end