+++ spec/dummy_environments/sqlite/dummy1/models/complex_thing.rb
@@ -1,0 +1,4 @@
+# This is serialized in the Post.serialized_thing attribute.
+#
+class ComplexThing
+end