+++ spec/dummy_environments/sqlite/dummy1/models/user/admin.rb
@@ -1,0 +1,3 @@
+require_relative '../user'
+class User::Admin < User
+end