+++ spec/dummy_environments/sqlite/dummy1/models/user/jerk.rb
@@ -1,0 +1,3 @@
+require_relative '../user'
+class User::Jerk < User
+end