+++ test/test_helper.rb
@@ -5,7 +5,6 @@ require 'shoulda'
 require 'iron_fixture_extractor'
 # TEMP, COMMENT OUT IF NEEDED
 require 'sqlite3'
-#require 'ruby-debug'
 require 'fe_test_env'
 ActiveRecord::Migration.verbose = false
 FeTestEnv.reload